requirements.txt
google-genai
flask
python-dotenv
Pillow
spacy
sumy
pyjwt
requests